gui-track-info.py: Add script to show in a GUI the metadata from the song played in Spotify
Beside the metadata of the song played in Spotify, it also includes the album
of the current song.

Only for this commit the album cover is from a different song which is used as
an example. In later commits this will change and the actual album cover of the
song will be shown.